Split celery-specific handlers into separate module
